Address review feedback. Changes: - Some style changes - TF_COMPILE_LIBRARY define controls exports from the tensorflow.dll.   If TF_COMPILE_LIBRARY is defined we compile tensorflow, if undefined   we compile a user op. - On windows user ops have now a .dll extension. The resource loader will   take care of finding it. - pywrap_tensorflow_internal.lib that is needed to build a user op on windows   comes with the wheel
